refactor: extract mailing logic from utils into utils/mailing.py
